Auto-import completed classes
This patch uses the CompilationUnitTree of the file being processed for a
completion to build a Set of imports for the current file and to get the path to
the file. It then uses the list of imports to determine whether the completion
candidate has already been imported. If not, the AddImport class is used to add
to the additionalTextEdits field on the CompletionItem.

Some small refactoring was done to pull useful class and package name extractors
out of JavaCompilerService.java.